Move price to global
