+++ code/client/src/integration/RequestPayment.jsx
@@ -16,7 +16,7 @@ const { Title, Paragraph } = Typography
 const RequestPayment = ({ caller, callback, amount, dest, from }) => {
   dest = util.safeNormalizedAddress(dest)
   const balances = useSelector(state => state.balance)
-  const price = useSelector(state => state.wallet.price)
+  const price = useSelector(state => state.global.price)
   const [selectedAddress, setSelectedAddress] = useState({})
   const { formatted: amountFormatted, fiatFormatted: amountFiatFormatted } = util.computeBalance(amount, price)
