+++ code/client/src/pages/List.jsx
@@ -32,7 +32,7 @@ const WalletCard = ({ wallet }) => {
   const oneAddress = getAddress(address).bech32
   const dispatch = useDispatch()
   const walletBalance = useSelector(state => state.balance[address] || {})
-  const price = useSelector(state => state.wallet.price)
+  const price = useSelector(state => state.global.price)
   const { formatted, fiatFormatted } = util.computeBalance(walletBalance.balance || 0, price)
   const walletOutdated = util.isWalletOutdated(wallet)
 
@@ -82,7 +82,7 @@ const List = () => {
   const { isMobile } = useWindowDimensions()
   const wallets = useSelector(state => state.wallet.wallets)
   const balances = useSelector(state => state.balance)
-  const price = useSelector(state => state.wallet.price)
+  const price = useSelector(state => state.global.price)
   const network = useSelector(state => state.wallet.network)
   const dispatch = useDispatch()
   const totalBalance = Object.keys(balances)
