+++ code/client/src/pages/Show/Call.jsx
@@ -51,7 +51,7 @@ const Call = ({
   const { resetWorker, recoverRandomness } = useRandomWorker()
 
   const balances = useSelector(state => state.balance)
-  const price = useSelector(state => state.wallet.price)
+  const price = useSelector(state => state.global.price)
   const { balance, formatted } = util.computeBalance(balances[address].balance || 0, price)
 
   const [transferTo, setTransferTo] = useState({ value: prefillDest || '', label: prefillDest ? util.oneAddress(prefillDest) : '' })
