+++ code/client/src/pages/Unwrap.jsx
@@ -160,7 +160,7 @@ const Unwrap = () => {
 
   const dispatch = useDispatch()
   const { isMobile } = useWindowDimensions()
-  const price = useSelector(state => state.wallet.price)
+  const price = useSelector(state => state.global.price)
   const network = useSelector(state => state.wallet.network)
   const wallets = useSelector(state => state.wallet.wallets)
   const balances = useSelector(state => state.balance)
