+++ code/client/src/state/modules/global/actions.js
@@ -10,6 +10,9 @@ const deleteKnownAddress = createAction('DELETE_KNOWN_ADDRESS')
 const setFetchStatus = createAction('SET_FETCH_STATUS')
 const setError = createAction('SET_ERROR')
 
+const fetchPrice = createAction('FETCH_PRICE')
+const fetchPriceSuccess = createAction('FETCH_PRICE_SUCCESS')
+
 export default {
   setDev,
 
@@ -17,4 +20,7 @@ export default {
   setKnownAddress,
   setError,
   setFetchStatus,
+
+  fetchPrice,
+  fetchPriceSuccess,
 }
